优化页面，增加权限认证

+++ springboot/src/main/java/com/ams/springboot/controller/RoleController.java
@@ -1,10 +1,13 @@
 package com.ams.springboot.controller;
 
 
+import com.ams.springboot.common.Constants;
 import com.ams.springboot.common.Result;
 import com.ams.springboot.entity.Am;
 import com.ams.springboot.entity.Role;
+import com.ams.springboot.entity.User;
 import com.ams.springboot.service.IRoleService;
+import com.ams.springboot.utils.TokenUtils;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import org.springframework.web.bind.annotation.*;
@@ -21,13 +24,23 @@ public class RoleController {
     //增加修改同一个方法
     @PostMapping
     public Result save(@RequestBody Role role) {
-        return Result.success(roleService.saveOrUpdate(role));
+        User user = TokenUtils.getCurrentUser();
+        if (isPower(user)){
+            return Result.success(roleService.saveOrUpdate(role));
+        }else {
+            return Result.error(Constants.CODE_401,"当前用户权限不足！");
+        }
     }
 
     //根据id删除记录
     @DeleteMapping("/{id}")
     public Result delete(@PathVariable Integer id) {
-        return Result.success(roleService.removeById(id));
+        User user = TokenUtils.getCurrentUser();
+        if (isPower(user)){
+            return Result.success(roleService.removeById(id));
+        }else {
+            return Result.error(Constants.CODE_401,"当前用户权限不足！");
+        }
     }
 
     //查询整个列表
@@ -57,5 +70,14 @@ public class RoleController {
 //            User currentUser = TokenUtils.getCurrentUser();
 //            System.out.println("获取当前用户信息==================" + currentUser.getNickname());
         return Result.success(roleService.page(new Page<>(pageNum, pageSize),queryWrapper));
+    }
+
+    //进行权限验证
+    public  boolean isPower(User user){
+        if (user.getSysroleid()==0 || user.getSysroleid()==1 || user.getSysroleid()==2){
+            return true;
+        }else {
+            return false;
+        }
     }
 }
